Address PR comments

+++ cosmobot_deep_learning/custom_metrics.py
@@ -57,11 +57,7 @@ def get_fraction_outside_error_threshold_fn(
         )
         count_total = tf.size(y_true)
 
-        # Cast to float so that the division calculation returns a float\
-        fraction_outside = tf.math.divide(
-            tf.cast(count_outside_error_threshold, tf.float32),
-            tf.cast(count_total, tf.float32),
-        )
+        fraction_outside = tf.math.divide(count_outside_error_threshold, count_total)
         return fraction_outside
 
     fn_name = _get_fraction_outside_error_threshold_fn_name(error_threshold_mg_l)
